+++ packages/block-editor/src/hooks/index.js
@@ -12,6 +12,6 @@ import './font-size';
 import './border-color';
 import './layout';
 
-export { useCustomSides } from './spacing';
+export { useCustomSides } from './dimensions';
 export { getBorderClassesAndStyles, useBorderProps } from './use-border-props';
 export { getColorClassesAndStyles, useColorProps } from './use-color-props';
