+++ packages/block-editor/src/hooks/style.js
@@ -37,7 +37,7 @@ import {
 	TYPOGRAPHY_SUPPORT_KEY,
 	TYPOGRAPHY_SUPPORT_KEYS,
 } from './typography';
-import { SPACING_SUPPORT_KEY, SpacingPanel } from './spacing';
+import { SPACING_SUPPORT_KEY, DimensionsPanel } from './dimensions';
 import useDisplayBlockControls from '../components/use-display-block-controls';
 
 const styleSupportKeys = [
@@ -225,7 +225,7 @@ export const withBlockControls = createHigherOrderComponent(
 						<TypographyPanel { ...props } />
 						<BorderPanel { ...props } />
 						<ColorEdit { ...props } />
-						<SpacingPanel { ...props } />
+						<DimensionsPanel { ...props } />
 					</>
 				) }
 				<BlockEdit { ...props } />
